+++ core.ts
@@ -1,4 +1,5 @@
-//import './manual_typings/globals.d.ts'
+import './manual_typings/globals.d.ts'
+
 export * from './src/facade/facade';
 export * from './src/core/di';
 export { Attr, Input, Output, HostBinding, HostListener, Directive, Component } from './src/core/directives';
