+++ src/core/di/forward_ref.ts
@@ -1,4 +1,4 @@
-import {Type, stringify, isFunction} from '../facade/lang';
+import {Type, stringify, isFunction} from '../../facade/lang';
 
 /**
  * An interface that a function passed into {@link forwardRef} has to implement.
