+++ src/core/directives/decorators.ts
@@ -4,7 +4,7 @@ import {
   TypeDecorator
 } from '../util/decorators';
 
-import {Type} from '../facade/lang';
+import {Type} from '../../facade/lang';
 
 import {
   ContentChildrenMetadata,
