+++ src/core/directives/directive_provider.ts
@@ -1,7 +1,7 @@
 import {DirectiveResolver} from "../linker/directive_resolver";
-import {Type,assign,isBlank,isPresent,isFunction,noop,resolveDirectiveNameFromSelector} from "../facade/lang";
-import {StringWrapper} from "../facade/primitives";
-import {StringMapWrapper} from "../facade/collections";
+import {Type,assign,isBlank,isPresent,isFunction,noop,resolveDirectiveNameFromSelector} from "../../facade/lang";
+import {StringWrapper} from "../../facade/primitives";
+import {StringMapWrapper} from "../../facade/collections";
 import {hasLifecycleHook} from "../linker/directive_lifecycles_reflector";
 import {LifecycleHooks} from "../linker/directive_lifecycle_interfaces";
 import {DirectiveMetadata,ComponentMetadata,LegacyDirectiveDefinition} from "./metadata_directives";
