+++ src/core/pipes/pipe_interfaces.ts
