+++ src/facade/facade.ts
@@ -1,2 +1,4 @@
 // Public API for Facade
 // None :)
+
+export {Type} from './lang';
